HostScheduleHandler, HostVoteHandler 적용. (로직 미완)
