+++ src/server/handler/normal/ScheduleHandler.java
@@ -1,10 +1,8 @@
-package server.handler;
+package server.handler.normal;
 
 import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStreamReader;
 import java.io.PrintWriter;
-import java.net.Socket;
 import java.util.Map;
 
 public class ScheduleHandler extends Thread {
