+++ src/server/handler/normal/VoteHandler.java
@@ -1,11 +1,8 @@
-package server.handler;
+package server.handler.normal;
 
 import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStreamReader;
 import java.io.PrintWriter;
-import java.net.Socket;
-import java.nio.Buffer;
 import java.util.Map;
 
 public class VoteHandler extends Thread {
